requirements.txt
Flask==2.3.3
torch==2.10.0
torchvision==0.25.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76